More fixes to flight pathing:
PathNavigateFlying:
Flight pathing checks if it's stuck every 10 ticks now, and clears the path to be able to path again if so.

FlyPathFinder:
If a point on the path is air, the path finder checks if the NPC will run into any blocks above that point due to its height.